+++ src/main/java/it/smartcommunitylab/playandgo/engine/controller/PlayerController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -26,6 +27,7 @@ public class PlayerController extends PlayAndGoController {
 	public Player addPlayer(
 			@RequestBody Player player,
 			HttpServletRequest request) throws Exception {
+		checkAdminRole(request);
 		return playerManager.addPlayer(player);
 	}
 	
@@ -40,6 +42,7 @@ public class PlayerController extends PlayAndGoController {
 	public Player getPlayer(
 			@PathVariable String playerId,
 			HttpServletRequest request) throws Exception {
+		checkAdminRole(request);
 		return playerManager.getPlayer(playerId);
 	}
 	
@@ -47,6 +50,7 @@ public class PlayerController extends PlayAndGoController {
 	public Player deletePlayer(
 			@PathVariable String playerId,
 			HttpServletRequest request) throws Exception {
+		checkAdminRole(request);
 		Player player = playerManager.deletePlayer(playerId);
 		//TODO delete other data
 		return player;
@@ -56,6 +60,14 @@ public class PlayerController extends PlayAndGoController {
 	public Player getProfile(
 			HttpServletRequest request) throws Exception {
 		return getCurrentPlayer(request);
+	}
+	
+	@PutMapping("/api/player/profile")
+	public Player updateProfile(
+			@RequestBody Player player,
+			HttpServletRequest request) throws Exception {
+		Player currentPlayer = getCurrentPlayer(request);
+		return playerManager.updatePlayer(player, currentPlayer.getPlayerId());
 	}
 
 }
